# apps/api/src/main/java/fr/epita/assistants/ping/presentation/api/request/RegisterRequest.java
package fr.epita.assistants.ping.presentation.api.request;

public class RegisterRequest {
  public String username;
  public String email;
  public String password;
  public String confirmPassword;
}