# apps/api/src/main/java/fr/epita/assistants/ping/presentation/api/response/LoginResponse.java
package fr.epita.assistants.ping.presentation.api.response;

public class LoginResponse {
  public String token;
  public UserResponse user;

  public LoginResponse(String token) {
    this.token = token;
  }

  public LoginResponse(String token, UserResponse user) {
    this.token = token;
    this.user = user;
  }
}